Create SuggestingAnActivity.java
it tells you the activities that can be done according to the air temperature.
If the temperature is less than 5, he recommends "Skiing".
If the temperature is between 5 and 15, he recommends the "Cinema" event.Dec.
If the temperature is between 15 and 25 Dec, he recommends the "Picnic" event.
If the temperature is greater than 25, he recommends the "Swimming" activity.